Refactoring loading private key into utils class
