+++ lib/mcapi/crypto/legacy-crypto.js
@@ -18,7 +18,7 @@ function LegacyCrypto(config) {
   this.encryptionCertificate = readPublicCertificate(config.encryptionCertificate);
 
   // Load private key (for decryption)
-  this.privateKey = getPrivateKey(config);
+  this.privateKey = utils.getPrivateKey(config);
 
   this.encoding = config.dataEncoding;
 
@@ -168,19 +168,6 @@ function generateSecretKey(algorithm, size, digest) {
 /**
  * @private
  */
-function loadPrivateKey(privateKeyPath) {
-  const privateKeyContent = fs.readFileSync(privateKeyPath, 'binary');
-
-  if (!privateKeyContent || privateKeyContent.length <= 1) {
-    throw new Error('Private key content not valid');
-  }
-
-  return forge.pki.privateKeyFromAsn1(forge.asn1.fromDer(privateKeyContent));
-}
-
-/**
- * @private
- */
 function readPublicCertificate(publicCertificatePath) {
   const certificateContent = fs.readFileSync(publicCertificatePath);
   if (!certificateContent || certificateContent.length <= 1) {
@@ -188,95 +175,6 @@ function readPublicCertificate(publicCertificatePath) {
   }
   return forge.pki.certificateFromPem(certificateContent);
 }
-
-/**
- * @private
- */
-function getPrivateKey(config) {
-  if (config.privateKey) {
-    return loadPrivateKey(config.privateKey);
-  } else if (config.keyStore) {
-    if (config.keyStore.includes(".p12")) {
-      return getPrivateKey12(config.keyStore, config.keyStoreAlias, config.keyStorePassword);
-    }
-    if (config.keyStore.includes(".pem")) {
-      return getPrivateKeyPem(config.keyStore);
-    }
-    if (config.keyStore.includes(".der")) {
-      return getPrivateKeyDer(config.keyStore);
-    }
-  }
-  return null;
-}
-
-/**
- * @private
- */
-function getPrivateKey12(p12Path, alias, password) {
-  const p12Content = fs.readFileSync(p12Path, 'binary');
-
-  if (!p12Content || p12Content.length <= 1) {
-    throw new Error('p12 keystore content is empty');
-  }
-
-  if (!utils.isSet(alias)) {
-    throw new Error('Key alias is not set');
-  }
-
-  if (!utils.isSet(password)) {
-    throw new Error('Keystore password is not set');
-  }
-
-  // Get asn1 from DER
-  const p12Asn1 = forge.asn1.fromDer(p12Content, false);
-
-  // Get p12 using the password
-  const p12 = forge.pkcs12.pkcs12FromAsn1(p12Asn1, false, password);
-
-  // Get Key from p12
-  const keyObj = p12.getBags({
-    friendlyName: alias,
-    bagType: forge.pki.oids.pkcs8ShroudedKeyBag
-  }).friendlyName[0];
-
-  if (!utils.isSet(keyObj)) {
-    throw new Error("No key found for alias [" + alias + "]");
-  }
-
-  return keyObj.key;
-}
-
-/**
- * Load a PKCS#1 PEM (starts with "-----BEGIN RSA PRIVATE KEY-----")
- * or a PKCS#8 PEM (starts with "-----BEGIN PRIVATE KEY-----")"
- */
-function getPrivateKeyPem(pemPath) {
-  let pemContent = fs.readFileSync(pemPath, 'binary');
-
-  if (!pemContent || pemContent.length <= 1) {
-    throw new Error('pem keystore content is empty');
-  }
-
-  pemContent = pemContent.replace("\n", "");
-  pemContent = pemContent.replace("\r\n", "");
-
-  return forge.pki.privateKeyFromPem(pemContent);
-}
-
-/**
- * Load a Binary DER-encoded PKCS#8
- */
-function getPrivateKeyDer(derPath) {
-  const derContent = fs.readFileSync(derPath, 'binary');
-
-  if (!derContent || derContent.length <= 1) {
-    throw new Error('der keystore content is empty');
-  }
-
-  const pkeyAsn1 = forge.asn1.fromDer(derContent);
-  return forge.pki.privateKeyFromAsn1(pkeyAsn1);
-}
-
 
 /**
  * @private
