+++ lib/mcapi/utils/utils.js
@@ -1,4 +1,5 @@
 const forge = require('node-forge');
+const fs = require('fs');
 
 /**
  * Utils module
@@ -12,6 +13,9 @@ const forge = require('node-forge');
  * @returns {boolean} True if set, false otherwise
  */
 module.exports.isSet = function (value) {
+  return typeof value !== "undefined" && value !== null && value.length !== 0;
+};
+isSet = function (value) {
   return typeof value !== "undefined" && value !== null && value.length !== 0;
 };
 
@@ -139,3 +143,100 @@ function deleteRoot(obj, paths, properties){
     });
   }
 }
+
+// module.exports.isNode10 = function(){
+//   const majorVersion = parseInt(parseInt(process.version.substring(1,3)));
+//   if(majorVersion < 10){
+//     throw new Error("Only node version 10+ supported");
+//   }
+//   if(majorVersion === 10){
+//     return true;
+//   }
+//   return false;
+// };
+
+module.exports.getPrivateKey = function(config){
+  if (config.privateKey) {
+    return loadPrivateKey(config.privateKey);
+  } else if (config.keyStore) {
+    if (config.keyStore.includes(".p12")) {
+      return getPrivateKey12(config.keyStore, config.keyStoreAlias, config.keyStorePassword);
+    }
+    if (config.keyStore.includes(".pem")) {
+      return getPrivateKeyPem(config.keyStore);
+    }
+    if (config.keyStore.includes(".der")) {
+      return getPrivateKeyDer(config.keyStore);
+    }
+  }
+  return null;
+};
+
+function getPrivateKey12(p12Path, alias, password) {
+  const p12Content = fs.readFileSync(p12Path, 'binary');
+
+  if (!p12Content || p12Content.length <= 1) {
+    throw new Error('p12 keystore content is empty');
+  }
+
+  if (!isSet(alias)) {
+    throw new Error('Key alias is not set');
+  }
+
+  if (!isSet(password)) {
+    throw new Error('Keystore password is not set');
+  }
+
+  // Get asn1 from DER
+  const p12Asn1 = forge.asn1.fromDer(p12Content, false);
+
+  // Get p12 using the password
+  const p12 = forge.pkcs12.pkcs12FromAsn1(p12Asn1, false, password);
+
+  // Get Key from p12
+  const keyObj = p12.getBags({
+    friendlyName: alias,
+    bagType: forge.pki.oids.pkcs8ShroudedKeyBag
+  }).friendlyName[0];
+
+  if (!isSet(keyObj)) {
+    throw new Error("No key found for alias [" + alias + "]");
+  }
+
+  return keyObj.key;
+}
+
+function getPrivateKeyPem(pemPath) {
+  let pemContent = fs.readFileSync(pemPath, 'binary');
+
+  if (!pemContent || pemContent.length <= 1) {
+    throw new Error('pem keystore content is empty');
+  }
+
+  pemContent = pemContent.replace("\n", "");
+  pemContent = pemContent.replace("\r\n", "");
+
+  return forge.pki.privateKeyFromPem(pemContent);
+}
+
+function getPrivateKeyDer(derPath) {
+  const derContent = fs.readFileSync(derPath, 'binary');
+
+  if (!derContent || derContent.length <= 1) {
+    throw new Error('der keystore content is empty');
+  }
+
+  const pkeyAsn1 = forge.asn1.fromDer(derContent);
+  return forge.pki.privateKeyFromAsn1(pkeyAsn1);
+}
+
+function loadPrivateKey(privateKeyPath) {
+  const privateKeyContent = fs.readFileSync(privateKeyPath, 'binary');
+
+  if (!privateKeyContent || privateKeyContent.length <= 1) {
+    throw new Error('Private key content not valid');
+  }
+
+  return forge.pki.privateKeyFromAsn1(forge.asn1.fromDer(privateKeyContent));
+}
+
